moved scrapers to /cinemas
