+++ scripts/cinemas/luxCinema.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import ICAL from 'ical.js';
-import type { CinemaShowing, FilmShowing } from '../src/types';
+import type { CinemaShowing, FilmShowing } from '../../src/types';
 
 async function fetchAndParseICS(
   url: string,
