+++ scripts/cinemas/theSpaceCinema.ts
@@ -1,4 +1,4 @@
-import type { CinemaShowing } from '../src/types';
+import type { CinemaShowing } from '../../src/types';
 
 export async function scraper(cinema: number = 1012): Promise<CinemaShowing[]> {
   // TODO - extend to every the space cinema
